Same code but more Composer friendly.
