--- imageLimits.php
@@ -1,15 +1,0 @@
-<?php
-
-//These are the limits for query images:
-//for SingleIR
-define("SINGLEIR_MAX_FILE_SIZE", 500);			//KBytes
-define("SINGLEIR_MIN_DIMENSION", 100);			//pix
-define("SINGLEIR_MIN_IMAGE_SURFACE", 0.05);		//Mpix
-define("SINGLEIR_MAX_IMAGE_SURFACE", 0.31);		//Mpix
-	
-//for MultipleIR
-define("MULTIIR_MAX_FILE_SIZE", 3500);			//KBytes
-define("MULTIIR_MIN_DIMENSION", 100);			//pix
-define("MULTIIR_MIN_IMAGE_SURFACE", 0.1);		//Mpix
-define("MULTIIR_MAX_IMAGE_SURFACE", 5.1);		//Mpix
-
